requirements.txt
flask
yt-dlp
openai-whisper
google-generativeai
python-dotenv
python-docx
gunicorn